requirements.txt
# GAIC Detector Web GUI - Requirements

# Core web frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio==4.44.0

# Image processing
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# ML frameworks (for model inference)
torch==2.1.0
torchvision==0.16.0

# Visualization
matplotlib==3.8.2

# HTTP client
requests==2.31.0
httpx==0.25.2

# Data validation
pydantic==2.5.0

# Optional: LLM integration
openai==1.3.0

# Optional: Additional model dependencies
# timm==0.9.12  # For vision transformers
# einops==0.7.0  # For tensor operations

# Development tools (optional)
python-multipart==0.0.6
python-dotenv==1.0.0
